refactor worldcloud
